Refactor publish pipeline around Git tags (GEA-12975)
Publishing to the NuGet registry is now triggered via the pushing of Git
tags instead of requiring the use of the one `release` branch. This
enables flexibility in publishing prerelease versions outside of the
`release` branch and shaves one step off of the manual release process
(merging to `release`).

Instructions on how to publish a new version going forward have been
added to `CONTRIBUTING.md`.